DEP: deprecate fs.retrieve and fs.commands
 - move all stateful logic into filestore.api
 - modify tests to import from fs.api
