+++ filestore/file_writers.py
@@ -11,7 +11,7 @@ import os
 import os.path as op
 import datetime
 
-import filestore.commands as fsc
+import filestore.api as fsc
 
 logger = logging.getLogger(__name__)
