+++ filestore/test/test_handlers.py
@@ -10,7 +10,7 @@ import uuid
 from filestore.api import (insert_resource, insert_datum, retrieve,
                            register_handler, deregister_handler)
 
-import filestore.retrieve as fsr
+from  filestore.api import handler_context
 from filestore.utils.testing import fs_setup, fs_teardown
 from filestore.handlers import AreaDetectorHDF5Handler
 from filestore.handlers import DummyAreaDetectorHandler
@@ -70,7 +70,7 @@ class test_np_FW(_with_file):
             insert_datum(resource_id, datum_id, dict(frame_no=i))
 
     def test_retrival(self):
-        with fsr.handler_context({'npy_FRAMEWISE': NpyFrameWise}):
+        with handler_context({'npy_FRAMEWISE': NpyFrameWise}):
             for i, datum_id in enumerate(self.datum_ids):
                 data = retrieve(datum_id)
                 known_data = i * np.ones((9, 8))
@@ -164,7 +164,7 @@ class test_maps_hdf5(_with_file):
     def test_maps_spectrum_round_trip(self):
         sn = np.sin(self.th)
 
-        with fsr.handler_context({'hdf5_maps': HDFM}):
+        with handler_context({'hdf5_maps': HDFM}):
             for eid, sc in zip(self.eids_spectrum, self.scale):
                 print(eid)
                 data = retrieve(eid)
@@ -172,7 +172,7 @@ class test_maps_hdf5(_with_file):
 
     def test_maps_plane_round_trip(self):
         base = self.scale.reshape(self.N, self.M)
-        with fsr.handler_context({'hdf5_planes': HDFE}):
+        with handler_context({'hdf5_planes': HDFE}):
             for eid, v in zip(self.eids_planes, np.sin(self.th)):
                 data = retrieve(eid)
                 assert_array_equal(data, base * v)
